scripts/data_collector/crypto/requirement.txt
loguru
fire
requests
numpy
pandas
tqdm
lxml
python-binance